# [Merged by Bors] - Optimise `per_epoch_processing` low-hanging-fruit