--- a/consensus/state_processing/src/per_epoch_processing/altair/rewards_and_penalties.rs
+++ b/consensus/state_processing/src/per_epoch_processing/altair/rewards_and_penalties.rs
@@ -7,7 +7,7 @@ use types::consts::altair::{
 use types::{BeaconState, ChainSpec, EthSpec};
 
 use crate::common::{
-    altair::{get_base_reward_optimised, get_base_reward_per_increment},
+    altair::{get_base_reward, BaseRewardPerIncrement},
     decrease_balance, increase_balance,
 };
 use crate::per_epoch_processing::{Delta, Error};
@@ -70,10 +70,10 @@ pub fn get_flag_index_deltas<T: EthSpec>(
     let unslashed_participating_increments =
         unslashed_participating_balance.safe_div(spec.effective_balance_increment)?;
     let active_increments = total_active_balance.safe_div(spec.effective_balance_increment)?;
-    let base_reward_per_increment = get_base_reward_per_increment(total_active_balance, spec)?;
+    let base_reward_per_increment = BaseRewardPerIncrement::new(total_active_balance, spec)?;
 
     for &index in participation_cache.eligible_validator_indices() {
-        let base_reward = get_base_reward_optimised(state, base_reward_per_increment, index, spec)?;
+        let base_reward = get_base_reward(state, index, base_reward_per_increment, spec)?;
         let mut delta = Delta::default();
 
         if unslashed_participating_indices.contains(index as usize)? {