@michaelsproul 2022-06-09
Nice! :tada: 

Regarding optimising the other base reward calls, we could consider caching the whole `base_reward` for each validator on the `ParticipationCache`. This is the approach I took on the `tree-states` branch when I noticed base reward calculations taking significant time, and for `tree-states` had the added advantage of reducing the number of expensive tree iterations. The relevant struct is called `ValidatorInfo` on `tree-states`, and also contains a bunch of other fields which probably aren't worth caching in the non-tree case:

https://github.com/sigp/lighthouse/blob/a3e396cfddf5fd5b3d35c5b05ed333ba00c8dca2/consensus/state_processing/src/per_epoch_processing/altair/participation_cache.rs#L133-L142

https://github.com/sigp/lighthouse/blob/a3e396cfddf5fd5b3d35c5b05ed333ba00c8dca2/consensus/state_processing/src/per_epoch_processing/altair/participation_cache.rs#L155-L159

That's a larger and more invasive change however, and seeing as this change combined with the Vec map already yields enormous speed-ups, maybe we take the simple win and merge this almost as-is :heart_eyes: